# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agro API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/policies": {
            "get": {
                "tags": [
                    "Policies"
                ],
                "summary": "Obtener la política",
                "description": "Retorna la política con su imagen asociada.",
                "operationId": "9e98360ee35089ef4fc6e1ad320e0c4d",
                "responses": {
                    "200": {
                        "description": "Éxito: Retorna la política",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Política de privacidad"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Esta es la política..."
                                        },
                                        "image": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://example.com/image.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Política no encontrada"
                    }
                }
            }
        },
        "/api/policies/{idPolicies}": {
            "put": {
                "tags": [
                    "Policies"
                ],
                "summary": "Actualizar la política",
                "description": "Actualiza la información de la política y su imagen.",
                "operationId": "473a686b7a283e960254c6aa4d811dbc",
                "parameters": [
                    {
                        "name": "idPolicies",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "aboutValue"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Política actualizada"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nueva descripción de la política."
                                    },
                                    "aboutValue": {
                                        "type": "string",
                                        "example": "Información adicional sobre la política."
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "byte",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANS..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Política actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Policy updated successfully"
                                        },
                                        "policy": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Política actualizada"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Nueva descripción..."
                                                },
                                                "aboutValue": {
                                                    "type": "string",
                                                    "example": "Información adicional actualizada."
                                                },
                                                "image": {
                                                    "properties": {
                                                        "url": {
                                                            "type": "string",
                                                            "example": "https://example.com/image.jpg"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor al guardar la imagen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error saving image"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}